fixed wasm_timespec

+++ runtime/libc/libc_backing.c
@@ -516,7 +516,7 @@ u32 wasm_geteuid() {
 
 #define SYS_GET_TIME 228
 struct wasm_time_spec {
-    u32 sec;
+    u64 sec;
     u32 nanosec;
 };
